Nsenene

Documentation for Nsenene.

Nsenene.spherical.density

Nsenene.spherical.total_mass

Nsenene.spherical.total_masses

Nsenene.spherical.total_masses

Nsenene.spherical.density — Method

density(profile::SphericalProfile, m)

Calculate the total mass density of profile with particle masses m.

Nsenene.spherical.total_mass — Method

total_mass(profile::SphericalProfile, m)

Calculate the total mass contained in profile, with particle masses m. Returns a scalar.

Examples

import Nsenene.spherical: SphericalProfile, total_mass

n = 3
m = ones(1, n)

p = SphericalProfile(1000, 3.0, n)

M = total_mass(p, m)

# output

0.0

Nsenene.spherical.total_masses — Method

total_masses(r, psi, m)

Calculate the mass contained by each field with radial coordinates r and particle masses m.

Examples

import Nsenene.spherical: total_masses

resol = 1000
nfields = 2

m = ones(1, nfields)
r = range(0, 10, resol)
psi = zeros(resol, nfields)

# Ball of radius 1
psi[r.<1, 1] .= 1

M = total_masses(psi, r, m)

@assert isapprox(M[1], 4/3 * π * 1^3, atol=1e-1)
M[2]

# output

0.0

Nsenene.spherical.total_masses — Method

total_masses(profile::SphericalProfile, m)

Calculate the mass contained by each field in profile, with particle masses m. Returns an array of masses.

Examples

import Nsenene.spherical: SphericalProfile, total_masses

n = 3
m = ones(1, n)

p = SphericalProfile(1000, 3.0, n)

# Ball of radius 1
p.psi[:, 1] .= p.r .< 1
p.psi[:, 2] .= p.r .< 2

M = total_masses(p, m)

@assert isapprox(M[1], 4/3 * π * 1^3, rtol=1e-2)
@assert isapprox(M[2], 4/3 * π * 2^3, rtol=1e-2)
M[3]

# output

0.0